% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{jamiecheck}
\alias{jamiecheck}
\title{Checking a dataframe}
\usage{
jamiecheck(df, ...)
}
\arguments{
\item{df}{dataframe}

\item{...}{more dataframes to check at once}
}
\description{
This function simply prints the dimensions of the dataframe, the first 6
rows of the dataframe and the names of the variables in the dataframe.
}
\examples{
jamiecheck(mtcars)
jamiecheck(mtcars, iris)
}
